requirements.txt
google-cloud-translate==2.0.0
lambda-decorators
six